requirements.txt
python-telegram-bot==20.6
Flask==2.3.3
pymongo==4.5.0
PyPDF2==3.0.1
Pillow==10.0.0
reportlab==4.0.4
cryptography==41.0.3
